analysis_service/requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
google-genai
python-dotenv
pydantic
requests
httpx
